requirements.txt
Django==4.1
djangorestframework==3.14
drf-yasg==1.21.7
django-filter==2.4.0
python-dateutil==2.9.0.post0
